Separate inventory upload functions
